IKASAN-1283 - fixing test after Merging 2.0.x-builderFactory-refactor
